Split out test that relies on compat
