+++ debug/test/browser/debug.test.js
@@ -10,7 +10,6 @@ import * as PropTypes from 'prop-types';
 
 // eslint-disable-next-line no-duplicate-imports
 import { resetPropWarnings } from 'preact/debug';
-import { forwardRef } from 'preact/compat';
 
 const h = createElement;
 /** @jsx createElement */
@@ -543,33 +542,6 @@ describe('debug', () => {
 					/^Failed prop type: Invalid prop `text` of type `number` supplied to `Foo`, expected `string`\.\n {2}in Foo \(at (.*)\/debug\/test\/browser\/debug\.test\.js:[0-9]+\)$/m
 				)
 			);
-		});
-
-		it('should not fail if ref is passed to comp wrapped in forwardRef', () => {
-			// This test ensures compat with airbnb/prop-types-exact, mui exact prop types util, etc.
-
-			const Foo = forwardRef(function Foo(props, ref) {
-				return <h1 ref={ref}>{props.text}</h1>;
-			});
-
-			Foo.propTypes = {
-				text: PropTypes.string.isRequired,
-				ref(props) {
-					if ('ref' in props) {
-						throw new Error(
-							'ref should not be passed to prop-types valiation!'
-						);
-					}
-				}
-			};
-
-			const ref = createRef();
-
-			render(<Foo ref={ref} text="123" />, scratch);
-
-			expect(console.error).not.been.called;
-
-			expect(ref.current).to.not.be.undefined;
 		});
 
 		it('should only log a given prop type error once', () => {
